refactor quotes into service and renderer
